Move gac_io.slice_channel to utils
To prevent pulling in h5py as extra dependency in satpy.
